chore: renderMapContent 메소드를 utils로 옮김
